requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Marshmallow
PyJWT
marshmallow-sqlalchemy
pandas
flask_restful

gunicorn
flask
pymysql
openpyxl
flask_cors
flask_sqlalchemy

flask_marshmallow
marshmallow-sqlalchemy
flask_jwt_extended
flask_mail
flask_bcrypt
flask_restful
flask_script
flask_migrate
flask_login
flask_admin
flask_wtf
Flask-Login 
Flask-Session
python-decouple
pymysql
mysql-connector-python
